Refatorando código
