+++ src/main/java/br/com/API/projetoGame/consoles/ConsolesApplication.java
@@ -1,0 +1,13 @@
+package br.com.API.projetoGame.consoles;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ConsolesApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ConsolesApplication.class, args);
+    }
+
+}
