+++ src/main/java/br/com/API/projetoGame/consoles/domains/enums/ModoDesempenho.java
@@ -1,0 +1,5 @@
+package br.com.API.projetoGame.consoles.domains.enums;
+
+public enum ModoDesempenho {
+    MODO_ECONOMIA_ENERGIA, MODO_PADRAO, MODO_DESEMPENHO
+}
