+++ src/main/java/br/com/API/projetoGame/consoles/repository/NintendoRepository.java
@@ -1,0 +1,9 @@
+package br.com.API.projetoGame.consoles.repository;
+
+import br.com.API.projetoGame.consoles.domains.NintendoDomain;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface NintendoRepository extends JpaRepository<NintendoDomain, Long> {
+}
