--- src/main/java/br/com/projetoGame/consoles/ConsolesApplication.java
@@ -1,13 +1,0 @@
-package br.com.projetoGame.consoles;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ConsolesApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ConsolesApplication.class, args);
-    }
-
-}
