--- src/main/java/br/com/projetoGame/consoles/models/enums/ModoDesempenho.java
@@ -1,5 +1,0 @@
-package br.com.projetoGame.consoles.models.enums;
-
-public enum ModoDesempenho {
-    MODO_ECONOMIA_ENERGIA, MODO_PADRAO, MODO_DESEMPENHO
-}
