--- src/main/java/br/com/projetoGame/consoles/repository/JogadorRepository.java
@@ -1,11 +1,0 @@
-package br.com.projetoGame.consoles.repository;
-
-import br.com.projetoGame.consoles.models.Jogador;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface JogadorRepository extends JpaRepository<Jogador, Long> {
-
-    List<Jogador> findByNome(String nome);
-}
