--- src/main/java/br/com/projetoGame/consoles/repository/JogosRepository.java
@@ -1,8 +1,0 @@
-package br.com.projetoGame.consoles.repository;
-
-import br.com.projetoGame.consoles.models.Jogos;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JogosRepository extends JpaRepository<Jogos, Long> {
-
-}
