+++ src/test/java/br/com/API/projetoGame/consoles/ConsolesApplicationTests.java
@@ -1,0 +1,13 @@
+package br.com.API.projetoGame.consoles;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ConsolesApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
