--- src/test/java/br/com/projetoGame/consoles/ConsolesApplicationTests.java
@@ -1,13 +1,0 @@
-package br.com.projetoGame.consoles;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ConsolesApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
